fix feedback from PR
